rename test classes to differ each other
