+++ app-jee-jsp/src/test/java/org/keycloak/quickstart/ArquillianJeeJspTest.java
@@ -21,6 +21,7 @@ import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.OperateOnDeployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.drone.api.annotation.Drone;
+import org.jboss.arquillian.graphene.Graphene;
 import org.jboss.arquillian.graphene.page.Page;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
@@ -30,39 +31,48 @@ import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.asset.EmptyAsset;
 import org.jboss.shrinkwrap.api.asset.StringAsset;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
+import org.jboss.shrinkwrap.resolver.api.maven.Maven;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.keycloak.quickstart.appjee.Controller;
 import org.keycloak.quickstart.page.IndexPage;
 import org.keycloak.quickstart.page.LoginPage;
-import org.keycloak.quickstart.page.ProfilePage;
-import org.keycloak.quickstart.profilejee.Controller;
 import org.keycloak.test.TestsHelper;
 import org.keycloak.test.builders.ClientBuilder;
+import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.support.ui.ExpectedConditions;
 
 import java.io.File;
 import java.io.IOException;
+import java.net.MalformedURLException;
 import java.net.URL;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
+import static org.keycloak.quickstart.page.IndexPage.MESSAGE_ADMIN;
+import static org.keycloak.quickstart.page.IndexPage.MESSAGE_PUBLIC;
+import static org.keycloak.quickstart.page.IndexPage.MESSAGE_SECURED;
+import static org.keycloak.quickstart.page.IndexPage.UNAUTHORIZED;
 import static org.keycloak.test.TestsHelper.createClient;
 import static org.keycloak.test.TestsHelper.deleteRealm;
 import static org.keycloak.test.TestsHelper.importTestRealm;
 import static org.keycloak.test.builders.ClientBuilder.AccessType.BEARER_ONLY;
+import static org.keycloak.test.builders.ClientBuilder.AccessType.PUBLIC;
 
 /**
  * @author <a href="mailto:bruno@abstractj.org">Bruno Oliveira</a>
  */
 @RunWith(Arquillian.class)
 @RunAsClient
-public class ArquillianTest {
+public class ArquillianJeeJspTest {
 
     private static final String WEBAPP_SRC = "src/main/webapp";
-    private static final String APP_NAME = "app-profile-saml";
+    private static final String APP_NAME = "app-jsp";
     private static final String APP_SERVICE = "service-jaxrs";
+    private static final String ROOT_URL = "http://127.0.0.1:8080/app-jsp";
 
     @Page
     private IndexPage indexPage;
@@ -70,19 +80,15 @@ public class ArquillianTest {
     @Page
     private LoginPage loginPage;
 
-    @Page
-    private ProfilePage profilePage;
-
     static {
         try {
             importTestRealm("admin", "admin", "/quickstart-realm.json");
-            createClient(ClientBuilder.create(APP_SERVICE).accessType(BEARER_ONLY));
         } catch (IOException e) {
             e.printStackTrace();
         }
     }
 
-    @Deployment(name = APP_SERVICE, order = 1, testable = false)
+    @Deployment(name= APP_SERVICE, order = 1, testable = false)
     public static Archive<?> createTestArchive1() throws IOException {
         return ShrinkWrap.createFromZipFile(WebArchive.class,
                 new File("../service-jee-jaxrs/target/service.war"))
@@ -91,15 +97,21 @@ public class ArquillianTest {
                                 ClientBuilder.create(APP_SERVICE).accessType(BEARER_ONLY))), "keycloak.json");
     }
 
-    @Deployment(name = APP_NAME, order = 2, testable = false)
+    @Deployment(name= APP_NAME, order = 2, testable = false)
     public static Archive<?> createTestArchive2() throws IOException {
-        return ShrinkWrap.create(WebArchive.class, "app-profile-saml.war")
+        File[] files = Maven.resolver().loadPomFromFile("pom.xml")
+                    .importRuntimeDependencies().resolve().withTransitivity().asFile();
+
+        return ShrinkWrap.create(WebArchive.class,  "app-jsp.war")
                 .addPackages(true, Filters.exclude(".*Test.*"), Controller.class.getPackage())
+                .addAsLibraries(files)
                 .addAsWebResource(new File(WEBAPP_SRC, "index.jsp"))
-                .addAsWebResource(new File(WEBAPP_SRC, "profile.jsp"))
+                .addAsWebResource(new File(WEBAPP_SRC, "protected.jsp"))
                 .addAsWebResource(new File(WEBAPP_SRC, "styles.css"))
                 .addAsWebInfResource(EmptyAsset.INSTANCE, "beans.xml")
-                .addAsWebInfResource(new File("src/test/resources", "keycloak-saml.xml"))
+                .addAsWebInfResource(new StringAsset(createClient(ClientBuilder.create(APP_NAME)
+                        .rootUrl(ROOT_URL)
+                        .accessType(PUBLIC))), "keycloak.json")
                 .setWebXML(new File("src/main/webapp", "WEB-INF/web.xml"));
     }
 
@@ -121,28 +133,58 @@ public class ArquillianTest {
     }
 
     @Test
-    public void testLogin() throws InterruptedException {
+    public void testSecuredResource() throws InterruptedException {
+        try {
+            indexPage.clickSecured();
+            assertTrue(Graphene.waitGui().until(ExpectedConditions.textToBePresentInElementLocated(By.className("error"), UNAUTHORIZED)));
+        } catch (Exception e) {
+            fail("Should display an error message");
+        }
+    }
+
+    @Test
+    public void testAdminResource() {
+        try {
+            indexPage.clickAdmin();
+            assertTrue(Graphene.waitGui().until(ExpectedConditions.textToBePresentInElementLocated(By.className("error"), UNAUTHORIZED)));
+        } catch (Exception e) {
+            fail("Should display an error message");
+        }
+    }
+
+    @Test
+    public void testPublicResource() {
+        try {
+            indexPage.clickPublic();
+            assertTrue(Graphene.waitGui().until(ExpectedConditions.textToBePresentInElementLocated(By.className("message"), MESSAGE_PUBLIC)));
+        } catch (Exception e) {
+            fail("Should display an error message");
+        }
+    }
+
+    @Test
+    public void testAdminWithAuthAndRole() throws MalformedURLException, InterruptedException {
         try {
             indexPage.clickLogin();
-            loginPage.login("alice", "password");
-            assertEquals(profilePage.getUsername(), "alice");
-            profilePage.clickLogout();
+            loginPage.login("test-admin", "password");
+            indexPage.clickAdmin();
+            assertTrue(Graphene.waitGui().until(ExpectedConditions.textToBePresentInElementLocated(By.className("message"), MESSAGE_ADMIN)));
+            indexPage.clickLogout();
         } catch (Exception e) {
             fail("Should display logged in user");
         }
     }
 
     @Test
-    public void testAccessAccountManagement() {
+    public void testUserWithAuthAndRole() throws MalformedURLException, InterruptedException {
         try {
             indexPage.clickLogin();
             loginPage.login("alice", "password");
-            profilePage.clickAccount();
-            assertEquals("Keycloak Account Management", webDriver.getTitle());
+            indexPage.clickSecured();
+            assertTrue(Graphene.waitGui().until(ExpectedConditions.textToBePresentInElementLocated(By.className("message"), MESSAGE_SECURED)));
+            indexPage.clickLogout();
         } catch (Exception e) {
-            fail("Should display account management page");
+            fail("Should display logged in user");
         }
     }
-
-
 }
