+++ app-profile-saml-jee-jsp/src/test/java/org/keycloak/quickstart/ArquillianProfileSamlJeeJspTest.java
@@ -15,16 +15,19 @@
  * limitations under the License.
  */
 
+package org.keycloak.quickstart;
+
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.OperateOnDeployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.drone.api.annotation.Drone;
-import org.jboss.arquillian.graphene.Graphene;
 import org.jboss.arquillian.graphene.page.Page;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.shrinkwrap.api.Archive;
+import org.jboss.shrinkwrap.api.Filters;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
+import org.jboss.shrinkwrap.api.asset.EmptyAsset;
 import org.jboss.shrinkwrap.api.asset.StringAsset;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.AfterClass;
@@ -33,40 +36,33 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.keycloak.quickstart.page.IndexPage;
 import org.keycloak.quickstart.page.LoginPage;
+import org.keycloak.quickstart.page.ProfilePage;
+import org.keycloak.quickstart.profilejee.Controller;
 import org.keycloak.test.TestsHelper;
 import org.keycloak.test.builders.ClientBuilder;
-import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.support.ui.ExpectedConditions;
 
 import java.io.File;
 import java.io.IOException;
-import java.net.MalformedURLException;
 import java.net.URL;
 
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
-import static org.keycloak.quickstart.page.IndexPage.MESSAGE_ADMIN;
-import static org.keycloak.quickstart.page.IndexPage.MESSAGE_PUBLIC;
-import static org.keycloak.quickstart.page.IndexPage.MESSAGE_SECURED;
-import static org.keycloak.quickstart.page.IndexPage.UNAUTHORIZED;
 import static org.keycloak.test.TestsHelper.createClient;
 import static org.keycloak.test.TestsHelper.deleteRealm;
 import static org.keycloak.test.TestsHelper.importTestRealm;
 import static org.keycloak.test.builders.ClientBuilder.AccessType.BEARER_ONLY;
-import static org.keycloak.test.builders.ClientBuilder.AccessType.PUBLIC;
 
 /**
  * @author <a href="mailto:bruno@abstractj.org">Bruno Oliveira</a>
  */
 @RunWith(Arquillian.class)
 @RunAsClient
-public class ArquillianTest {
+public class ArquillianProfileSamlJeeJspTest {
 
     private static final String WEBAPP_SRC = "src/main/webapp";
-    private static final String APP_NAME = "app-html5";
+    private static final String APP_NAME = "app-profile-saml";
     private static final String APP_SERVICE = "service-jaxrs";
-    private static final String ROOT_URL = "http://127.0.0.1:8080/app-html5";
 
     @Page
     private IndexPage indexPage;
@@ -74,9 +70,13 @@ public class ArquillianTest {
     @Page
     private LoginPage loginPage;
 
+    @Page
+    private ProfilePage profilePage;
+
     static {
         try {
             importTestRealm("admin", "admin", "/quickstart-realm.json");
+            createClient(ClientBuilder.create(APP_SERVICE).accessType(BEARER_ONLY));
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -93,14 +93,14 @@ public class ArquillianTest {
 
     @Deployment(name = APP_NAME, order = 2, testable = false)
     public static Archive<?> createTestArchive2() throws IOException {
-        return ShrinkWrap.create(WebArchive.class, "app-html5.war")
-                .addAsWebResource(new File(WEBAPP_SRC, "app.js"))
-                .addAsWebResource(new File(WEBAPP_SRC, "index.html"))
-                .addAsWebResource(new File(WEBAPP_SRC, "keycloak.js"))
+        return ShrinkWrap.create(WebArchive.class, "app-profile-saml.war")
+                .addPackages(true, Filters.exclude(".*Test.*"), Controller.class.getPackage())
+                .addAsWebResource(new File(WEBAPP_SRC, "index.jsp"))
+                .addAsWebResource(new File(WEBAPP_SRC, "profile.jsp"))
                 .addAsWebResource(new File(WEBAPP_SRC, "styles.css"))
-                .addAsWebResource(new StringAsset(createClient(ClientBuilder.create(APP_NAME)
-                        .rootUrl(ROOT_URL)
-                        .accessType(PUBLIC))), "keycloak.json");
+                .addAsWebInfResource(EmptyAsset.INSTANCE, "beans.xml")
+                .addAsWebInfResource(new File("src/test/resources", "keycloak-saml.xml"))
+                .setWebXML(new File("src/main/webapp", "WEB-INF/web.xml"));
     }
 
     @Drone
@@ -111,8 +111,8 @@ public class ArquillianTest {
     private URL contextRoot;
 
     @AfterClass
-    public static void cleanUp() throws IOException {
-        deleteRealm("admin", "admin", TestsHelper.testRealm);
+    public static void cleanUp() throws IOException{
+        deleteRealm("admin","admin",TestsHelper.testRealm);
     }
 
     @Before
@@ -121,58 +121,28 @@ public class ArquillianTest {
     }
 
     @Test
-    public void testSecuredResource() throws InterruptedException {
-        try {
-            indexPage.clickSecured();
-            assertTrue(Graphene.waitGui().until(ExpectedConditions.textToBePresentInElementLocated(By.className("error"), UNAUTHORIZED)));
-        } catch (Exception e) {
-            fail("Should display an error message");
-        }
-    }
-
-    @Test
-    public void testAdminResource() {
-        try {
-            indexPage.clickAdmin();
-            assertTrue(Graphene.waitGui().until(ExpectedConditions.textToBePresentInElementLocated(By.className("error"), UNAUTHORIZED)));
-        } catch (Exception e) {
-            fail("Should display an error message");
-        }
-    }
-
-    @Test
-    public void testPublicResource() {
-        try {
-            indexPage.clickPublic();
-            assertTrue(Graphene.waitGui().until(ExpectedConditions.textToBePresentInElementLocated(By.className("message"), MESSAGE_PUBLIC)));
-        } catch (Exception e) {
-            fail("Should display an error message");
-        }
-    }
-
-    @Test
-    public void testAdminWithAuthAndRole() throws MalformedURLException, InterruptedException {
-        try {
-            indexPage.clickLogin();
-            loginPage.login("test-admin", "password");
-            indexPage.clickAdmin();
-            assertTrue(Graphene.waitGui().until(ExpectedConditions.textToBePresentInElementLocated(By.className("message"), MESSAGE_ADMIN)));
-            indexPage.clickLogout();
-        } catch (Exception e) {
-            fail("Should display logged in user");
-        }
-    }
-
-    @Test
-    public void testUserWithAuthAndRole() throws MalformedURLException, InterruptedException {
+    public void testLogin() throws InterruptedException {
         try {
             indexPage.clickLogin();
             loginPage.login("alice", "password");
-            indexPage.clickSecured();
-            assertTrue(Graphene.waitGui().until(ExpectedConditions.textToBePresentInElementLocated(By.className("message"), MESSAGE_SECURED)));
-            indexPage.clickLogout();
+            assertEquals(profilePage.getUsername(), "alice");
+            profilePage.clickLogout();
         } catch (Exception e) {
             fail("Should display logged in user");
         }
     }
+
+    @Test
+    public void testAccessAccountManagement() {
+        try {
+            indexPage.clickLogin();
+            loginPage.login("alice", "password");
+            profilePage.clickAccount();
+            assertEquals("Keycloak Account Management", webDriver.getTitle());
+        } catch (Exception e) {
+            fail("Should display account management page");
+        }
+    }
+
+
 }
